Refactor pistelaskuriSlice
