+++ src/main/app/src/store/reducers/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
-import { pistelaskuriSlice } from '../../components/laskuri/hooks';
 import appSlice from './appSlice';
 import hakutulosSlice from './hakutulosSlice';
 import koulutusSlice from './koulutusSlice';
 import oppilaitosSlice from './oppilaitosSlice';
+import pistelaskuriSlice from './pistelaskuriSlice';
 
 export default combineReducers({
   koulutus: koulutusSlice,
